pyproject.toml
[tool.poetry]
name = "tinyurl"
version = "0.1.0"
description = ""
authors = ["dmpetraytis <daniil.petraytis@yandex.ru>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
sqlmodel = "^0.0.22"
fastapi = {extras = ["standard"], version = "^0.115.5"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
pre-commit = "^4.0.1"
pydantic-settings = "2.1.0"
aioredis = "^2.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-asyncio = "^0.18.3"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
